transplanting Alexa work into main wordle-web directory
